Changed from double to long double for greater precision. May later templatize all functions so we can freely switch between types. Pondering

+++ include/wacfrac/bla.hpp
@@ -7,23 +7,23 @@ namespace wacfrac {
 
 class bivariate_linear_approximator {
     public:
-    bivariate_linear_approximator(double epsilon, std::complex<double> max_dc, const std::vector<std::complex<double>>& ref, std::size_t first_level = 0);
-    bivariate_linear_approximator(double initial_epsilon, double tolerance, const std::vector<std::complex<double>>& probes, std::complex<double> max_dc, const std::vector<std::complex<double>>& ref, std::size_t first_level = 0);
-    auto escape_approximate(std::complex<double> dc) const -> std::tuple<std::complex<double>, std::size_t, std::size_t>;
+    bivariate_linear_approximator(double epsilon, std::complex<long double> max_dc, const std::vector<std::complex<long double>>& ref, std::size_t first_level = 0);
+    bivariate_linear_approximator(double initial_epsilon, double tolerance, const std::vector<std::complex<long double>>& probes, std::complex<long double> max_dc, const std::vector<std::complex<long double>>& ref, std::size_t first_level = 0);
+    auto escape_approximate(std::complex<long double> dc) const -> std::tuple<std::complex<long double>, std::size_t, std::size_t>;
 
     private:
-    bivariate_linear_approximator(const std::vector<std::complex<double>>& ref, std::size_t first_level);
+    bivariate_linear_approximator(const std::vector<std::complex<long double>>& ref, std::size_t first_level);
 
     struct bla {
-        std::complex<double> a, b;
-        double r;
-        auto is_valid(std::complex<double> dzm) const -> bool;
-        auto approximate_dzn(std::complex<double> dzm, std::complex<double> dc) const -> std::complex<double>;
+        std::complex<long double> a, b;
+        long double r;
+        auto is_valid(std::complex<long double> dzm) const -> bool;
+        auto approximate_dzn(std::complex<long double> dzm, std::complex<long double> dc) const -> std::complex<long double>;
     };
-    auto compute_bla(double epsilon, std::complex<double> max_dc, std::size_t m, std::size_t n) const -> bla;
-    auto compute_blas(double epsilon, std::complex<double> max_dc) -> void;
-    auto merge_blas(std::complex<double> max_dc, const bla& x, const bla& y) const -> bla;
-    auto compute_zn(std::complex<double> dc, std::complex<double> dzm, std::size_t m) const -> std::optional<std::pair<std::complex<double>, std::size_t>>;
+    auto compute_bla(double epsilon, std::complex<long double> max_dc, std::size_t m, std::size_t n) const -> bla;
+    auto compute_blas(double epsilon, std::complex<long double> max_dc) -> void;
+    auto merge_blas(std::complex<long double> max_dc, const bla& x, const bla& y) const -> bla;
+    auto compute_zn(std::complex<long double> dc, std::complex<long double> dzm, std::size_t m) const -> std::optional<std::pair<std::complex<long double>, std::size_t>>;
 
     struct column_info {
         std::size_t first;
@@ -33,7 +33,7 @@ class bivariate_linear_approximator {
     auto bla_at(std::size_t m, std::size_t level) const -> const bla*;
     auto bla_at(std::size_t m, std::size_t level) -> bla*;
 
-    const std::vector<std::complex<double>>& _ref;
+    const std::vector<std::complex<long double>>& _ref;
     std::size_t _first_level;
     std::size_t _last_level;
     std::vector<bla> _blas;
